Pin libffi subproject to a particular commit.
The 'centricular/libffi' project has introduced a change that makes
Meson 0.47.0 the minimum supported version.

Since our older CentOS 6.9 and Ubuntu 14.04 containers run Python 3.4,
moving to a newer version of Meson is not really an option at this time.

To work around this problem, this commit pins the libffi subproject to a
commit just prior to the version change.